모델도 통합 하나로 정리
