+++ src/search_spotify.py
@@ -4,10 +4,7 @@ root_path = os.path.join(os.path.dirname(os.path.abspath(__file__)),'..')
 sys.path.append(root_path)
 
 from model.database import session_scope
-from model.spotify_artists import SpotifyArtists, SpotifyArtistsORM
-from model.spotify_albums import SpotifyAlbums, SpotifyAlbumsORM
-from model.spotify_tracks import SpotifyTracks, SpotifyTracksORM
-from model.spotify_search import SpotifySearch, SpotifySearchArtists
+import model.spotify_search as Spotify
 
 from src.get_token import update_token, return_token
 
@@ -27,22 +24,22 @@ if __name__ == '__main__':
     if response.status_code == 200 :
         responsed_data = response.json()
 
-        parsed_data = SpotifySearch(**responsed_data)
+        parsed_data = Spotify.Search(**responsed_data)
         artists = parsed_data.artists.items
         albums = parsed_data.albums.items
         tracks = parsed_data.tracks.items
 
         for entity in artists :
-            orm = SpotifyArtistsORM(entity)
+            orm = Spotify.ArtistsORM(entity)
             with session_scope() as session :
                 session.add(orm)
 
         for entity in albums :
-            orm = SpotifyAlbumsORM(entity)
+            orm = Spotify.AlbumsORM(entity)
             with session_scope() as session :
                 session.add(orm)
 
         for entity in tracks :
-            orm = SpotifyTracksORM(entity)
+            orm = Spotify.TracksORM(entity)
             with session_scope() as session :
                 session.add(orm)
